Update models to differentiate non-literal values
Previously, the models did not differentiate between `LiteralValue<*>`s
created from literal expressions and those created from other
expressions, e.g. mathematical operations, fields. This change makes
them do that. It is needed because `AnnotationItem.toSource()` formats
values differently depending on whether they were literal or
non-literal.

The signature file does not really support expressions, apart from a
few special cases like `(0.0/0.0)` for `DoubleValue.NaN`, etc. However,
it can use the different ways `AnnotationItem.toSource()` formats them
to differentiate where necessary to preserve the behavior when reading
from and then writing back to a signature file.

These changes required some updates to some `ValueExample`s to use the
correct `expectedValue` and to tweak them to work properly with
signature files. In a couple of places it was impossible for the test
to pass for both legacy and non-legacy tests. In that case the latter
was prioritised which meant that a few legacy tests had to be
baselined. That does not mean that the code is broken it just means
that these specific tests cannot work in both cases.

Bug: 399365344
Test: ./gradlew
Change-Id: I51f8d403924a461987282e0995a215e84f7217ef