Add tests for negative floats and a float expression
`AnnotationItem.toSource()` formats float expressions differently to
float literals, it uses a `F` suffix instead of `f`. It considers a
negative float to be a unary minus expression even when the float is
loaded from a jar where it is stored as a negative number.

This change adds tests to show that behavior.

Bug: 399365344
Test: ./gradlew
Change-Id: I4bbcf65c596fc8815a181efcafd38a64b37a0545