Support formatting non-literal floats with `F` suffix
The `AnnotationItem.toSource()` formats non-literal floatss using an
`F` suffix instead of `f`. Where a non-literal float is one that was
not represented as a literal in the source but instead was computed
from an expression. Historically, Psi has considered a negative float
to be a unary minus expression, even when loaded from a jar where it
is represented as a negative int.

This change adds a `nonLiteralInSource` property to `DefaultFloatValue`
to track whether it was a non-literal in the source (or negative) and
adds `ValueStringConfiguration.nonLiteralFloatSuffix` to set the suffix
to use for a non-literal float.

Bug: 399365344
Test: ./gradlew
Change-Id: If4fe9d3a817fde981f616736cfec9710fe556d52