Add nonLiteralInSource to createLiteralValue(...)
`AnnotationItem.toSource()` produces different output for some
`LiteralValue`s depending on whether they were represented as a literal
in the source, or a more complex expression. This change adds a
`nonLiteralInSource` to `createLiteralValue(...)` and wires it into the
infrastructure used to create `LiteralValue`s. This does nothing on its
own but is separated out as it is needed by a number of the following
changes. Doing that makes it easier to manage, e.g. reorganize, those
changes.

Bug: 354633349
Test: ./gradlew
Change-Id: I6461f3efa6a2459d1921b9aeb81d80d8330f7e6a