Add tests for negative ints with numeric types
Negative ints are treated specially by `AnnotationItem.toSource()` as
they are a unary expression and not a literal. This change adds tests
to show that is also true when they are used with other numeric types,
i.e. byte, double, float, long, short. The important part of the test
is that the `annotationToSource` is a hexadecimal number.

This reveals a minor difference between K1 and K2 in the handling of
a negative int when used with a long type. That will be fix in a
follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: I404f851575eed5881993561a85fcfabd4c4afcb4