Remove Expectations.attributeValue
Previously `Expectations.attributeValue` was used to specialize the
expectations for `AnnotationAttribute.legacyValue` (both for
`AnnotationAttributeValue.toSource()` and `...value()`). However,
`leagcyValue` was removed in a preceding change.

It could potentially be used to specialize the expectation for
`AnnotationAttribute.value` but that does not need specializing for a
particular use (that is pretty much the whole point of it).

So, this change removed `Expectations.attributeValue` as it is no
longer needed.

Bug: 354633349
Test: ./gradlew
Change-Id: I4a05f9ac22b57a6ca8cac707ae55299e2dde59ca