Refactor Settings.alwaysInlineFields
Previously, `Settings` had a `Boolean` property `alwaysInlineFields`
which controlled whether fields were always inlined or only inlined
when they were inaccessible. While that was sufficient for legacy Java
formatting it is not for Kotlin as legacy Kotlin formatting used in
`UAnnotationItem.toSource()` has slightly different criteria for
determining when to inline a field.

This change replaces the `Boolean` with an enum. A follow-up change
will extend that enum for the legacy Kotlin behavior.

Bug: 354633349
Test: ./gradlew
Change-Id: I198184da5bd086f0cb97b7fcfabc5047fd4861c8