Add test for Psi edge case in handling Float special values
Floats are stored in a class constant pool using their standard bit
representation, that includes the special floats, i.e. `NaN` and
`+/-Infinity`. In the source they cannot be represented as a literal
and so are represented using a division-by-zero expression, e.g.
`Float.NaN` is `(0.0f/0.0f)`. Psi emulates that behavior in its model
by representing it as a similar, but unfortunately, not identical
expression. It uses `(0.0f/0.0)` which evaluates to a `Double` not a
`Float`.

This change adds a test in `ParameterizedValueStringTest` to show the
String representations of such a value. Basically, they are treated as
if it was a float. This is needed to prevent a follow-up change from
changing that behavior.

Bug: 354633349
Test: ./gradlew
Change-Id: I0d30aed90d2a11405fc854c8eb4578f23f0a84db