TurbineValueFactory: Refactor determination of original source literal
Turbine has inconsistent behavior in how it represents negative numbers
from the source in its `Tree` model. Negative ints, e.g. `-2` and also
`- 2` are represented as `Tree.Literal`. However, negative floats, e.g.
`-2.0f` are represented as `Tree.Unary` with a `-` operator and a
`Tree.Literal` expression.

This change refactors the logic in `TurbineValueFactory` that tries to
determine the original source literal to handle both structures. This
is in preparation for the follow-up change that needs to treat float
literals similar to ints.

Bug: 354633349
Test: ./gradlew
Change-Id: Ic256eeb466b308a21db1e3cc2cd271057cd2fddf