Improve testing of expected FieldReferenceValues
Previously, `checkExpectedValue(...)` compared `FieldReferenceValue`s
just like any `Value` using their `equals(...)` method and by comparing
their `toString()` results. Neither checked the `FieldReferenceValue`'s
constant value.

This change expands the test to also check a `FieldReferenceValue`'s
constant value. That revealed a problem with the constant not correctly
recording that it should be treated as `nonLiteral` (because a field
reference is not a literal value). This change temporarily baselines
the newly failing tests and they will be fixed in a follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: I3a6b538677ccd192a145cd773632d5abca85724c