Stop Turbine optimizing field values
Previously, `TurbineValueFactory` would optimize the handling of field
values by only checking for `ConstantValue`s. That relied on Turbine
already having done the evaluation of the constants and avoided
checking for other value types. Unfortunately, when the field value
was a reference to a field that lost information about the original
type of the referenced field which was needed to replicate legacy
behavior.

This change removes that optimization and treats fields exactly like
annotations. That fixes a previously broken test.

Longer term, once users of Metalava have been migrated off the legacy
behavior the optimization may be able to be replaced, depending on the
canonical representation that is chosen to replace it.

Bug: 354633349
Test: ./gradlew
Change-Id: I350e8a154a8c1e77595ca1bde6d6b6f54ea72eeb