Find accessors of property items created with kotlin analysis API
This copies the existing behavior of propagating all getter annotations to a property, and propagating experimental annotations from properties to accessors.

Additionally, this maintains the behavior of using the inlined type for a value class property, to find the right psi accessors and avoid making a large change to AndroidX signature files for now.

There are a few cases where this doesn't link the accessors for K1 due to psi bugs, reflected in the test updates. These are value class extension properties and constructor property parameters with accessors renamed with JvmName.

Bug: 421200464
Test: ./gradlew test, usage in androidx (aosp/3636945)
Change-Id: I717e00faa066c34d68e7e377fcd1f040dd2916d6