CommonAnnotationItemTest: Compare with Values
Previously, `CommonAnnotationItemTest` would test annotations using
`toSource(...)` and `assertAttributeValue/s()`. The `toSource(...)`
method should not be part of the model tests and the latter involve
some complicated logic to map from the `AnnotationAttribute.value` to
primitive/String values.

This change replaces both with a simple comparison with a `Value`.
In preparation for removing `toSource(...)`.

Bug: 354633349
Test: ./gradlew
Change-Id: I9421ad12d280a706cd96f9517b3b87e434c8a5be