Use DefaultAnnotationItem instead of Psi/UAnnotationItem
Previously, `Psi/UAnnotationItem.create(...)` would create a
`Psi/UAnnotationItem`. The problem with that is neither of them
implemented `snapshot(Codebase)` correctly; they both overrode the
correct implementation in `DefaultAnnotationItem` with one that just
returned `this`. That broke a couple of tests when run against the
snapshot.

This change switches them to return `DefaultAnnotationItem` instead
which fixes the issue.

Bug: 354633349
Test: ./gradlew
Change-Id: I44d5d4ffdccb9737edda3b20590232b96c07fc6a