Add AnnotationFormatter abstraction around LegacyValueFormatter
Currently, `AnnotationItem.toSource()` uses a `LegacyValueFormatter`
directly. This change adds an `AnnotationFormatter` abstraction that
hides the `LegacyValueFormatter` and prepares for it to be replaced
with non-legacy implementations.

The `LegacyAnnotationFormatter` class was added to wrap a
`LegacyValueFormatter` rather than have the latter implement
`AnnotationFormatter` directly to simplify some of the follow-up
changes that will move some functionality out of
`LegacyValueFormatter`.

The `LegacyAnnotationFormatter` is returned by a function rather
than exposing the property as follow-up changes will parameterize the
creation by adding parameters to the function, e.g. `AnnotationTarget`.

Preparation for removing `AnnotationItem.toSource(...)`.

Bug: 354633349
Test: ./gradlew
Change-Id: I90e77b5f213ce9a4dc86537969c855da73278041