Add special AnnotationFormatter for stub files
Previously, the stub files used the legacy `AnnotationFormatter` to
maintain the existing behavior. However, there is no specific reason
for it to maintain that formatting as stub file content is not as
strictly controlled as the signature files. As long as it compiles a
true representation of the API surface and produces the correct
documentation (when necessary) the exact contents does not matter. e.g.
it does not matter if an integer is formatted as `-1` or `0xffffffff`.

This change adds a special `StubFormattter` that mostly replicates the
existing behavior without relying on the source representation, e.g.
the legacy formatter would format `-1` as `0xffffffff` because it was
represented as a unary expression in the source. However, the stub
formatter ignores that and just formats it as `-1`.

It does try and replicate the previous behavior as much as possible to
avoid unnecessary churn.

Bug: 354633349
Test: ./gradlew
Change-Id: I2f27c9723b61738c447810dbfe8fccafc18a9342