Add ValueStringConfiguration.inlineFieldReferenceChecker
Previously, `FieldReferenceValue.toValueString(...)` would always
return a string representation of the field reference. This change adds
`inlineFieldReferenceChecker` to allow the caller to control whether a
`FieldReferenceValue` is represented as a field reference or if the
constant value (if any) is inlined.

This is needed for use in a follow-up change to make sure that fields
which are not part of the API surface are not written into stub files.

Bug: 354633349
Test: ./gradlew
Change-Id: I8eb122f2d39ef6cc6793ad818fe3a9484555a1d5