Remove target parameter from AnnotationFormatter methods
Previously, the `AnnotationFormatter` methods took a `target` parameter
as that is what `AnnotationItem.toSource(...)` did. While that was
necessary for `toSource(...)` which was used for multiple different
`AnnotationTarget`s it is not necessary for `AnnotationFormatter` as a
separate instance can be created for each `AnnotationTarget` which will
make them simpler and more efficient.

This change removes the `target` parameter from those methods and
passes it into the factory method instead.

Bug: 354633349
Test: ./gradlew
Change-Id: If2dfff8e35bab83eff879c9de16176f7cca01445