Sort annotation attributes in stub files
Previously, stub files would output annotation attributes in the order
in which they were specified in the source file. That made the stub
files dependent on changes to the source files that did not affect the
API. This change causes it to sort the attributes before formatting
avoiding that dependency.

Bug: 354633349
Test: ./gradlew
Change-Id: Id85bd8d823a0b91f2a828503f0e7546e0481b5ba