Extract stubs specific copy of FieldItem.writeValueWithSemicolon(...)
Previously, `JavaStubWriter` would write a `FieldItem`'s value using
`FieldItem.writeValueWithSemicolon(...)`. That required the
`writeValueWithSemicolon(...)` method to provide a way to supply a
non-constant initializer for `static final` fields without a constant
value and return whether any value was written to allow for the
`JavaStubWriter` to use an initialization block instead. That
complicated both the calling code and `writeValueWithSemicolon(...)`.

This change copies `writeValueWithSemicolon(..)` to
`FieldItem.writeValue(...)` in `JavaStubWriter` and removes the stub
specific functionality from `writeValueWithSemicolon(..)`.

This duplicates a lot of formatting code but that will be cleaned up in
follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: Ib774049160eb19425a61c6ab8c2ad03291a6f334