Update PsiItemDocumentation creation to not require a PsiItem
Previously PsiItemDocumentation enforced that the Item it was being created for was a PsiItem, but the only reason that was needed was to use the attached PsiBasedCodebase. However, the creation already required specifying a PsiBasedCodebase, so it isn't necessary to get the codebase from the item.

This makes it so items created using the kotlin analysis api can be DefaultItems instead of PsiItems, but still use psi for documentation.

Bug: 407735992
Test: ./gradlew test

Change-Id: I44794ea567b63e8ca59ebc757aaebf76146c98f7