Implement isSamCompatibleOrKotlinLambda for all ParameterItems
This replaces the psi-specific definition of isSamCompatibleOrKotlinLambda with one on the base ParameterItem. It checks if a type is a lambda type or is an interface which is SAM compatible. This has the limitation (copied from the psi implementation) that if a type from the classpath is not known to be from kotlin source, there might be false-positives.

Bug: 421201575
Test: ./gradlew test
Change-Id: Ie08a012dccd173e72ed698a639f203babd42dc42